# src/express/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Memora API
  description: Memora openApi 3.0.1 specification
  version: 1.0.23
  contact:
    email: to@mail.io
servers:
  - description: Localhost
    url: http://localhost:3001
...
paths:
...
  /users/create:
    post:
      tags:
        - User
      summary: Creates a card
      operationId: createCardCtr
      description: |
        Creates a friendship connection between two users by their subs.

        Granted for a user with status: **GUEST**, **REGISTERED**
      security:
        - userLoginBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: email
                  example: example@gmail.com
                password:
                  type: string
                  description: password
                  example: pass123

      responses:
        '200':
          description: Connection status
          content:
            application/json:
              schema:
                type: object
                properties:
                  connection:
                    type: string
                    description: Connection status
                    example: completed
...
  /cards:
    get:
      tags:
        - Card
      summary: Returns cards
      operationId: getCardCtr
      description: |
        The profile will be provided according to the auth token provided by the user

        Granted for a user with status: **GUEST**, **REGISTERED**
      security:
        - userLoginBearerAuth: []
      responses:
        '200':
          description: Personal user profile
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    front:
                      type: string
                    back:
                      type: string
                    created_at:
                      type: string
  /cards/create:
    post:
      tags:
        - Card
      summary: Creates a card
      operationId: createCardCtr
      description: |
        Creates a friendship connection between two users by their subs.

        Granted for a user with status: **GUEST**, **REGISTERED**
      security:
        - userLoginBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                front:
                  type: string
                  description: front text
                  example: one
                back:
                  type: string
                  description: back text
                  example: two

      responses:
        '200':
          description: Connection status
          content:
            application/json:
              schema:
                type: object
                properties:
                  connection:
                    type: string
                    description: Connection status
                    example: completed
  /desks/create:
    post:
      tags:
        - Card
      summary: Creates a desk
      operationId: createDeskCtr
      description: |
        Creates a friendship connection between two users by their subs.

        Granted for a user with status: **GUEST**, **REGISTERED**
      security:
        - userLoginBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: desk title
                  example: desk title
                description:
                  type: string
                  description: desk description
                  example: desk description

      responses:
        '200':
          description: Connection status
          content:
            application/json:
              schema:
                type: object
                properties:
                  connection:
                    type: string
                    description: Connection status
                    example: completed
...
components:
...
  securitySchemes:
    userLoginBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT